Initial outline of Helix Trajectory Controller
This is a start for the subscriber/publisher node that will control Bebop trajectory.  

a few parts are inconplete.  The postion and velocity error in the normal/binormal directions is not done.  The controller is also not finished.  

The helix trajectory has been generated.  The code should properly read in odom telemetry from the bebop and be able to publish Twist commands to the vehicle.